% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fci.R
\name{plot.fcimodel}
\alias{plot.fcimodel}
\title{Plot a Financial Conditions Index Model}
\usage{
\method{plot}{fcimodel}(x, which = "fci", factor = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"fcimodel"} containing an estimated model.}

\item{which}{what to plot: \code{"fci"} for the index, \code{"loadings"} for
the time varying loadings of the financial variables, or \code{"sigma"} for
the measurement error variances.}

\item{factor}{an integer of the factor to plot.}

\item{...}{arguments passed forward to \code{\link[stats]{plot.ts}}.}
}
\value{
\code{NULL}, invisibly. Called for its side effect.
}
\description{
Plot a Financial Conditions Index Model
}
\examples{

data("koop")

model <- create_fcimodel(y = koop$macro[[96]], x = koop$financial, p = 4, n = 1)
model <- add_priors(model)
model <- add_posterior_coefficients(model)

plot(model)

}
